Fix innoverio.vscode-dbt-power-user Exasol errors
The innoverio.vscode-dbt-power-user VScode extension expects all errors
coming from DBT adapters to be raised as DBT errors. The dbt-exasol
adapter was raising Exasol DB errors as pyexasol's QueryError error type
and so this caused problems in the vscode extension.

This commit changes the error to the standard DBT DB error type.